only the fraction tracker uses the first argument passed to store(...), and the fraction tracker renders proper fractions (x/y where x and y are positive integers), so always pass a uint64 to store(...) (this change effectively reinstates commit 720837b). delete the "unitsDone" field of the Progress struct by changing the store(...) method of the fraction tracker to call Add() instead of Store(), so that the "current" field accumulates units of completed work. update some variable names and some comments in the Report(...) method.

+++ percent.go
@@ -8,6 +8,6 @@ type percentTracker struct {}
 //             to *Progress.state) per Report calls, but the very low-precision
 //             percentage rendered to the terminal does not change
 
-func (p *percentTracker) store(_ float64, _ string) {}
-func (p *percentTracker) load()  any                { return nil }
-func (p *percentTracker) value(_ any)       string  { return ""  }
+func (p *percentTracker) store(_ uint64, _ string) {}
+func (p *percentTracker) load()  any               { return nil }
+func (p *percentTracker) value(_ any)      string  { return ""  }
